bring upstream updates

+++ Dev/Dev2.Runtime.Services/ResourceCatalogImpl/ResourceDeleteProvider.cs
@@ -18,12 +18,12 @@ using Warewolf.Resource.Errors;
 
 namespace Dev2.Runtime.ResourceCatalogImpl
 {
-    internal class ResourceDeleteProvider : IResourceDeleteProvider
+    class ResourceDeleteProvider : IResourceDeleteProvider
     {
-        private readonly FileWrapper _dev2FileWrapper = new FileWrapper();
+        readonly FileWrapper _dev2FileWrapper = new FileWrapper();
         readonly IServerVersionRepository _serverVersionRepository;
-        private readonly IResourceCatalog _resourceCatalog;
-       
+        readonly IResourceCatalog _resourceCatalog;
+
 
         public ResourceDeleteProvider(IResourceCatalog resourceCatalog, IServerVersionRepository serverVersionRepository)
         {
@@ -53,7 +53,7 @@ namespace Dev2.Runtime.ResourceCatalogImpl
 
                 var workspaceResources = _resourceCatalog.GetResources(workspaceID);
                 var resources = _resourceCatalog.GetResourcesBasedOnType(type, workspaceResources, r => string.Equals(r.ResourceName, resourceName, StringComparison.InvariantCultureIgnoreCase));
-                Dictionary<int, ResourceCatalogResult> commands = new Dictionary<int, ResourceCatalogResult>()
+                var commands = new Dictionary<int, ResourceCatalogResult>()
                 {
                     {
                      0,ResourceCatalogResultBuilder.CreateNoMatchResult($"<Result>{type} '{resourceName}' was not found.</Result>")
@@ -101,12 +101,12 @@ namespace Dev2.Runtime.ResourceCatalogImpl
         
         }
 
-        private ResourceCatalogResult DeleteFromWorkspace(Guid workspaceID, Guid resourceID, string type, bool deleteVersions)
+        ResourceCatalogResult DeleteFromWorkspace(Guid workspaceID, Guid resourceID, string type, bool deleteVersions)
         {
             var @lock = Common.GetWorkspaceLock(workspaceID);
-            lock(@lock)
+            lock (@lock)
             {
-                if(resourceID == Guid.Empty || string.IsNullOrEmpty(type))
+                if (resourceID == Guid.Empty || string.IsNullOrEmpty(type))
                 {
                     throw new InvalidDataContractException(ErrorResource.ResourceNameAndTypeMissing);
                 }
@@ -115,7 +115,7 @@ namespace Dev2.Runtime.ResourceCatalogImpl
                 var resources = workspaceResources.FindAll(r => Equals(r.ResourceID, resourceID));
 
                 var commands = GetDeleteCommands(workspaceID, resourceID, type, deleteVersions, resources, workspaceResources);
-                if(commands.ContainsKey(resources.Count))
+                if (commands.ContainsKey(resources.Count))
                 {
                     var resourceCatalogResult = commands[resources.Count];
                     return resourceCatalogResult;
@@ -127,19 +127,17 @@ namespace Dev2.Runtime.ResourceCatalogImpl
         #endregion
 
         #region Private functions
-        private ResourceCatalogResult DeleteImpl(Guid workspaceID, IEnumerable<IResource> resources, List<IResource> workspaceResources, bool deleteVersions = true)
+        ResourceCatalogResult DeleteImpl(Guid workspaceID, IEnumerable<IResource> resources, List<IResource> workspaceResources, bool deleteVersions = true)
         {
 
-            IResource resource = resources.FirstOrDefault();
+            var resource = resources.FirstOrDefault();
 
-            if (workspaceID == Guid.Empty && deleteVersions)
+            if (workspaceID == Guid.Empty && deleteVersions && resource != null)
             {
-                if (resource != null)
-                {
-                    var explorerItems = _serverVersionRepository.GetVersions(resource.ResourceID);
-                    explorerItems?.ForEach(a => _serverVersionRepository.DeleteVersion(resource.ResourceID, a.VersionInfo.VersionNumber, resource.GetResourcePath(workspaceID)));
-                }
+                var explorerItems = _serverVersionRepository.GetVersions(resource.ResourceID);
+                explorerItems?.ForEach(a => _serverVersionRepository.DeleteVersion(resource.ResourceID, a.VersionInfo.VersionNumber, resource.GetResourcePath(workspaceID)));
             }
+
 
             workspaceResources.Remove(resource);
             if (resource != null && _dev2FileWrapper.Exists(resource.FilePath))
@@ -161,16 +159,13 @@ namespace Dev2.Runtime.ResourceCatalogImpl
                 };
                 UpdateDependantResourceWithCompileMessages(workspaceID, resource, messages);
             }
-            if (workspaceID == GlobalConstants.ServerWorkspaceID)
+            if (workspaceID == GlobalConstants.ServerWorkspaceID && resource != null)
             {
-                if (resource != null)
-                {
-                    ServiceActionRepo.Instance.RemoveFromCache(resource.ResourceID);
-                    ServerAuthorizationService.Instance.Remove(resource.ResourceID);
-                }
+                ServiceActionRepo.Instance.RemoveFromCache(resource.ResourceID);
+                ServerAuthorizationService.Instance.Remove(resource.ResourceID);
             }
             
-            ((ResourceCatalog)_resourceCatalog).RemoveFromResourceActivityCache(workspaceID, resource);
+            _resourceCatalog.RemoveFromResourceActivityCache(workspaceID, resource);
             return ResourceCatalogResultBuilder.CreateSuccessResult("Success");
         }
         void UpdateDependantResourceWithCompileMessages(Guid workspaceID, IResource resource, IList<ICompileMessageTO> messages)
@@ -201,9 +196,9 @@ namespace Dev2.Runtime.ResourceCatalogImpl
         }
 
 
-        private Dictionary<int, ResourceCatalogResult> GetDeleteCommands(Guid workspaceID, Guid resourceID, string type, bool deleteVersions, IEnumerable<IResource> resources, List<IResource> workspaceResources)
+        Dictionary<int, ResourceCatalogResult> GetDeleteCommands(Guid workspaceID, Guid resourceID, string type, bool deleteVersions, IEnumerable<IResource> resources, List<IResource> workspaceResources)
         {
-            Dictionary<int, ResourceCatalogResult> commands = new Dictionary<int, ResourceCatalogResult>()
+            var commands = new Dictionary<int, ResourceCatalogResult>()
             {
                 {
                     0,
@@ -215,6 +210,6 @@ namespace Dev2.Runtime.ResourceCatalogImpl
         }
 
         #endregion
-        
+
     }
 }
